+++ demo/src/elements/tabs.tsx
@@ -11,7 +11,7 @@ interface Props {
 }
 
 // todo rewrite with <wup-tabs />: https://www.w3.org/WAI/ARIA/apg/patterns/tabs/examples/tabs-automatic/
-export default function Tabs(props: Props): JSX.Element {
+export default function Tabs(props: Props): React.JSX.Element {
   const [current, setCurrent] = useState(0);
   const onClick = useCallback((e: React.MouseEvent) => {
     const ch = (e.currentTarget as HTMLElement).children;