+++ src/controls/selectMany.ts
@@ -35,6 +35,9 @@ declare global {
   interface HTMLElementTagNameMap {
     [tagName]: WUPSelectManyControl; // add element to document.createElement
   }
+}
+
+declare module "react" {
   namespace JSX {
     interface IntrinsicElements {
       /** Form-control with dropdown/combobox behavior