+++ src/formElement.ts
@@ -101,6 +101,9 @@ declare global {
   interface HTMLElementTagNameMap {
     [tagName]: WUPFormElement; // add element to document.createElement
   }
+}
+
+declare module "react" {
   namespace JSX {
     interface IntrinsicElements {
       /** Wrapper of FormHTMLElement that collect values from controls